Removed ExpenseContext and Added GetItems in Home.js
