+++ tracklah-frontend/components/index.js
@@ -1,10 +1,8 @@
 import { calculateTotal, calculateCategoryTotal } from "./CalculateTotal";
-import colorScale from "../styles/PieChartColorScale";
 import Card from "./card";
 
 export { 
     Card,
     calculateTotal, 
     calculateCategoryTotal, 
-    colorScale, 
     } ;
