+++ tracklah-frontend/contexts/index.js
@@ -1,4 +1,3 @@
-import { ExpenseContext } from "./ExpenseContext";
 import UserContext from "./UserContext";
 
-export {ExpenseContext, UserContext}
+export {UserContext};
