+++ tracklah-frontend/styles/index.js
@@ -2,10 +2,12 @@ import colors from "./colors";
 import styles from "./home-styles";
 import cardStyles from "./card-styles";
 import btnStyles from "./button-styles";
+import colorScale from "./piecolor-style";
 
 export {
     colors, 
     styles, 
     cardStyles, 
-    btnStyles
+    btnStyles,
+    colorScale,
 };
